Added Masternode Tool
These tools allow you to copy, edit, add, delete a masternode entry in the masternode.conf file from the GUI. It allows you to auto fill genkey and also auto fill already made, but unused masternode ouputs in a click of a button. This eliminates the need to use the debug console and edit the masternode.conf file when setting up and editing masternode configurations.